fix: rename filename of the clearErrors method
