+++ bigbluebutton-html5/imports/api/presentations/server/methods.js
@@ -4,7 +4,7 @@ import setPresentationRenderedInToast from './methods/setPresentationRenderedInT
 import setPresentation from './methods/setPresentation';
 import setPresentationDownloadable from './methods/setPresentationDownloadable';
 import exportPresentationToChat from './methods/exportPresentationToChat';
-import clearErrors from './methods/clearErros';
+import clearErrors from './methods/clearErrors';
 
 Meteor.methods({
   removePresentation,
