+++ bigbluebutton-html5/imports/api/presentations/server/methods/clearErrors.js
